Move the contents of enemy.py to sprites.py
